Move the DH pieces that were originally in the base spec into this
This radically simplifies the base specification.  The intent is
to avoid changing anything here, since we have deployed code.  The
net effect should be that the messages are identical, but the
description of the overall process should be a lot more direct
and proscriptive.  That should make this more usable all around.